requirements.txt
python-dotenv~=1.0.1
vosk~=0.3.44
pyaudio~=0.2.14
nltk~=3.9.1
google-cloud-aiplatform~=1.67.1
markdown~=3.7
requests~=2.32.3
msal~=1.31.0
msal-extensions~=1.2.0
protobuf~=5.28.2

# for testing whisper
openai-whisper
numpy
torch
